Carry the thread heap counts as std::size_t
hpx.thread_queue.init_threads_count and cached_threads_count end up in
reserve() calls on the thread heaps and in comparisons against their
size(), so std::size_t is what they are. They were held as std::uint64_t
in thread_queue_init_parameters, which MSVC reports as C4244 at the four
reserve() calls on a 32 bit target.

Hold them as std::size_t, and read them with get_entry_as<std::size_t>
rather than reading a std::int64_t into a std::uint64_t. The two
configuration reads next to these in create_pools, hpx.numa_sensitive
and hpx.max_background_threads, already use get_entry_as<std::size_t>.

The other counts in thread_queue_init_parameters stay std::int64_t.
Unlike these two they are compared against signed queue lengths rather
than used as sizes.

Refs #7576

+++ libs/core/schedulers/include/hpx/schedulers/thread_queue.hpp
@@ -1321,7 +1321,7 @@ namespace hpx::threads::policies {
                 "without changing the code here.");
 
             std::lock_guard<mutex_type> lk(mtx_);
-            for (std::uint64_t i = 0; i < parameters_.init_threads_count_; ++i)
+            for (std::size_t i = 0; i < parameters_.init_threads_count_; ++i)
             {
                 // We don't care about the init parameters since this thread
                 // will be rebound once it is actually used
